Parse imported types
